intelligence-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis[hiredis]==5.1.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
geoalchemy2==0.15.2
apscheduler==3.10.4
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
shapely==2.0.6
boto3>=1.34.0
ollama==0.4.4
